+++ apps/web/src/app/layouts/RootLayout.tsx
@@ -2,14 +2,17 @@ import { Outlet } from "react-router";
 import { LgpdModal } from "../components/LgpdModal";
 import { AuthProvider } from "../auth/AuthContext";
 import { ThemeProvider } from "../theme/ThemeContext";
+import { LanguageProvider } from "../i18n/LanguageContext";
 
 export function RootLayout() {
   return (
-    <ThemeProvider>
-      <AuthProvider>
-        <Outlet />
-        <LgpdModal />
-      </AuthProvider>
-    </ThemeProvider>
+    <LanguageProvider>
+      <ThemeProvider>
+        <AuthProvider>
+          <Outlet />
+          <LgpdModal />
+        </AuthProvider>
+      </ThemeProvider>
+    </LanguageProvider>
   );
 }